completed adding jwt token logic

+++ controllers/feed.js
@@ -1,8 +1,10 @@
 const fs = require("fs");
 const path = require("path");
 
-const { validationResult } = require("express-validator");
+const { validationResult } = require("express-validator/");
+
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -18,7 +20,7 @@ exports.getPosts = (req, res, next) => {
     })
     .then((posts) => {
       res.status(200).json({
-        message: "fetched post successfully",
+        message: "Fetched posts successfully.",
         posts: posts,
         totalItems: totalItems,
       });
@@ -34,31 +36,40 @@ exports.getPosts = (req, res, next) => {
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error("validation failed, entered data is incorrect");
-    error.statusCode = 442;
+    const error = new Error("Validation failed, entered data is incorrect.");
+    error.statusCode = 422;
     throw error;
   }
   if (!req.file) {
-    const error = new Error("no image provided");
+    const error = new Error("No image provided.");
     error.statusCode = 422;
     throw error;
   }
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
-  //Create post in db
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "abu" },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -74,14 +85,11 @@ exports.getPost = (req, res, next) => {
   Post.findById(postId)
     .then((post) => {
       if (!post) {
-        const error = new Error("Could not find post");
+        const error = new Error("Could not find post.");
         error.statusCode = 404;
         throw error;
       }
-      res.status(200).json({
-        message: "Post Fetched",
-        post: post,
-      });
+      res.status(200).json({ message: "Post fetched.", post: post });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -92,46 +100,46 @@ exports.getPost = (req, res, next) => {
 };
 
 exports.updatePost = (req, res, next) => {
+  const postId = req.params.postId;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error("validation failed, entered data is incorrect");
-    error.statusCode = 442;
+    const error = new Error("Validation failed, entered data is incorrect.");
+    error.statusCode = 422;
     throw error;
   }
-  const postId = req.params.postId;
   const title = req.body.title;
   const content = req.body.content;
-  let imageUrl = req.body.imageUrl;
-  console.log(imageUrl);
+  let imageUrl = req.body.image;
   if (req.file) {
     imageUrl = req.file.path;
   }
   if (!imageUrl) {
-    const error = new Error("no file picked");
+    const error = new Error("No file picked.");
     error.statusCode = 422;
     throw error;
   }
-
   Post.findById(postId)
     .then((post) => {
       if (!post) {
-        const error = new Error("could not find post");
+        const error = new Error("Could not find post.");
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
       post.title = title;
-      post.content = content;
       post.imageUrl = imageUrl;
+      post.content = content;
       return post.save();
     })
     .then((result) => {
-      res.status(200).json({
-        message: "post updated",
-        post: result,
-      });
+      res.status(200).json({ message: "Post updated!", post: result });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -143,23 +151,31 @@ exports.updatePost = (req, res, next) => {
 
 exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
-  console.log(postId);
-
   Post.findById(postId)
     .then((post) => {
-      console.log(post);
       if (!post) {
-        const error = new Error("could not find post");
+        const error = new Error("Could not find post.");
         error.statusCode = 404;
         throw error;
       }
-      //check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
-      res.status(200).json({ message: "deleted post!" });
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
+      res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
       if (!err.statusCode) {
